refactor history page
